Material/מטלות/maman6/requirements.txt
Flask==3.0.0
bcrypt==4.1.2
argon2-cffi==23.1.0
pyotp==2.9.0
requests==2.31.0
